refactor: Extracted functions (update-addon-package-json)
