+++ src/migration/ember-addon/steps/update-addon-package-json/index.ts
@@ -1,0 +1,4 @@
+export * from './update-dependencies.js';
+export * from './update-dev-dependencies.js';
+export * from './update-other-fields.js';
+export * from './update-scripts.js';
